chore(Flow): dedupe state viewer types with shared yt-types [#1752]
