+++ packages/ui/src/ui/pages/flow/flow-components/FlowStateViewer/types.ts
@@ -1,13 +1,12 @@
-export type FlowStateTarget = 'all' | 'key_state' | 'partition_state' | 'external_key_state';
+import type {
+    FlowDeleteStatesResponse,
+    FlowKeyColumn,
+    FlowReadStatesResponse,
+    FlowStateTarget,
+    GetPipelineStateData,
+} from '../../../../../shared/yt-types';
 
 export type FlowStateNameInputMode = 'declared-only' | 'suggested' | 'free-form';
-
-export type FlowKeyColumn = {
-    name: string;
-    type: string;
-    expression?: string;
-    required?: boolean;
-};
 
 export type FlowStateFiltersValue = {
     computationId?: string;
@@ -16,14 +15,6 @@ export type FlowStateFiltersValue = {
     stateName?: string;
     target: FlowStateTarget;
     limit: number;
-};
-
-export type FlowStateAccessBody = {
-    computation_id?: string;
-    partition_id?: string;
-    key?: unknown;
-    name?: string;
-    target?: FlowStateTarget;
 };
 
 export type FlowStateValidationErrorKey =
@@ -39,46 +30,6 @@ export type FlowStateValidationErrorKey =
 export type FlowStateValidationError = {
     errorKey: FlowStateValidationErrorKey;
     params?: Record<string, string>;
-};
-
-export type FlowAnnotatedInteger = {$type: 'int64' | 'uint64'; $value: string};
-
-export type FlowReadStatesBody = FlowStateAccessBody & {limit?: number};
-export type FlowDeleteStatesBody = FlowStateAccessBody & {force?: boolean; commit?: boolean};
-
-export type FlowKeyStateRow = {
-    computation_id: string;
-    key: unknown;
-    states: Record<string, unknown>;
-};
-
-export type FlowPartitionStateRow = {
-    computation_id?: string;
-    partition_id: string;
-    states: Record<string, unknown>;
-};
-
-export type FlowReadStatesResponse = {
-    key_states?: Array<FlowKeyStateRow>;
-    partition_states?: Array<FlowPartitionStateRow>;
-    external_key_states?: Array<FlowKeyStateRow>;
-    joined_external_key_states?: Array<FlowKeyStateRow>;
-    errors?: Array<string>;
-};
-
-export type FlowMatchedStatesBucket = {
-    total: number;
-    details?: Record<string, Record<string, number>>;
-};
-
-export type FlowDeleteStatesResponse = {
-    committed?: boolean;
-    matched_states?: {
-        key_states?: FlowMatchedStatesBucket;
-        partition_states?: FlowMatchedStatesBucket;
-        external_key_states?: FlowMatchedStatesBucket;
-    };
-    errors?: Array<string>;
 };
 
 export type FlowRowDeleteOutcome = {
@@ -102,15 +53,6 @@ export type FlowStateResultRow = {
     value: unknown;
 };
 
-export type FlowPipelineStateValue =
-    | 'Unknown'
-    | 'Stopped'
-    | 'Paused'
-    | 'Working'
-    | 'Draining'
-    | 'Pausing'
-    | 'Completed';
-
 export type FlowStateReadState = {
     revision: number;
     requestId: number;
@@ -131,7 +73,7 @@ export type FlowDeleteDialogState = {
     session: number;
     force: boolean;
     busy?: 'state' | 'preview' | 'delete';
-    pipelineState?: FlowPipelineStateValue;
+    pipelineState?: GetPipelineStateData;
     preview?: Array<FlowRowDeleteOutcome>;
     previewSnapshot?: FlowDeleteDialogSnapshot;
     committed?: Array<FlowRowDeleteOutcome>;
@@ -144,7 +86,7 @@ export type FlowDeleteDialogEvent =
     | {type: 'closed'; session: number}
     | {type: 'force-changed'; force: boolean}
     | {type: 'run-started'; commit: boolean}
-    | {type: 'pipeline-state-loaded'; session: number; pipelineState: FlowPipelineStateValue}
+    | {type: 'pipeline-state-loaded'; session: number; pipelineState: GetPipelineStateData}
     | {
           type: 'preview-loaded';
           session: number;
@@ -160,35 +102,6 @@ export type FlowDeleteDialogEvent =
     | {type: 'request-failed'; session: number; error: unknown};
 
 export type FlowPipelineSpecData = {spec?: unknown};
-
-export type FlowGroupBySchema =
-    | Array<FlowKeyColumn>
-    | {$attributes?: Record<string, unknown>; $value?: Array<FlowKeyColumn>};
-
-export type FlowYsonWrapped<T> = T | {$attributes?: Record<string, unknown>; $value?: T};
-
-export type FlowStateJoinSpec = {
-    key_schema_override?: FlowGroupBySchema;
-    key_provider_streams?: FlowYsonWrapped<Array<string>>;
-};
-
-export type FlowExternalStateJoinerSpec = {
-    join_on?: FlowYsonWrapped<FlowStateJoinSpec>;
-    parameters?: unknown;
-};
-
-export type FlowStaticSpecComputation = {
-    group_by_schema?: FlowGroupBySchema;
-    external_state_managers?: Record<string, unknown>;
-    external_state_joiners?: FlowYsonWrapped<
-        Record<string, FlowYsonWrapped<FlowExternalStateJoinerSpec>>
-    >;
-    heavy_hitters?: unknown;
-};
-
-export type FlowStaticSpec = {
-    computations?: Record<string, FlowStaticSpecComputation>;
-};
 
 export type FlowKeySchemaResolution = {
     keyColumns: Array<FlowKeyColumn>;
